+++ Tycoon.Backend.Application/Abstractions/IAppDb.cs
@@ -71,6 +71,7 @@ namespace Tycoon.Backend.Application.Abstractions
         DbSet<PlayerTransaction> PlayerTransactions { get; }
         DbSet<PlayerTransactionActor> PlayerTransactionActors { get; }
         DbSet<PlayerTransactionItem> PlayerTransactionItems { get; }
+        DbSet<PlayerPreferences> PlayerPreferences { get; }
 
         Task<int> SaveChangesAsync(CancellationToken ct = default);
         EntityEntry Entry(object entity);