+++ Tycoon.Backend.Infrastructure/Persistence/AppDb.cs
@@ -88,6 +88,7 @@ namespace Tycoon.Backend.Infrastructure.Persistence
         public DbSet<PlayerTransaction> PlayerTransactions => Set<PlayerTransaction>();
         public DbSet<PlayerTransactionActor> PlayerTransactionActors => Set<PlayerTransactionActor>();
         public DbSet<PlayerTransactionItem> PlayerTransactionItems => Set<PlayerTransactionItem>();
+        public DbSet<PlayerPreferences> PlayerPreferences => Set<PlayerPreferences>();
 
         protected override void OnModelCreating(ModelBuilder modelBuilder)
         {